sync.py can redownload all occurrences
If it detects that the number of occurrence records differs from the
number at ALA, Syncer will redownload the entire set of occurrences and
delete the local records that do not exist at ALA.

This is an expensive brute-force fallback, and it should only be
triggered very rarely when the more optimised update has failed. Also
gets triggered when ALA deletes records, because there is no other way
to find which records were deleted.