Added history.js so users can paste links to maps
Changes the url of the page without reloading the
page. So a user can type a bird into the 
autocomplete box, hit return, then copy and paste
the URL to someone else and that someone else will
see the same map as the sender. Currently sends
user to /species/map/X where X is the species id.